requirements.txt
APScheduler==3.11.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
feedparser==6.0.11
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.9
PyMySQL==1.1.1
schedule==1.2.2
sgmllib3k==1.0.0
tzdata==2025.2
tzlocal==5.3.1
Werkzeug==3.1.3
gunicorn==23.0.0
packaging==25.0
requests==2.32.3
Flask-Minify==0.46
